test(stylex-transform): add tests for media query canonicalization
