+++ crates/stylex-transform/tests/__swc_snapshots__/tests/transform_stylex_create_test/media_query_canonicalization.rs/authored_media_queries_are_canonicalized.js
@@ -1,0 +1,29 @@
+import _inject from "@stylexjs/stylex/lib/stylex-inject";
+var _inject2 = _inject;
+import * as stylex from '@stylexjs/stylex';
+_inject2({
+    ltr: ".x1s85apg{display:none}",
+    priority: 3000
+});
+_inject2({
+    ltr: "@media (min-width: 720px) and (max-height: 120px){.x1gcnmh1.x1gcnmh1{display:block}}",
+    priority: 3200
+});
+_inject2({
+    ltr: ".x1e2nbdu{color:red}",
+    priority: 3000
+});
+_inject2({
+    ltr: "@media (min-width: 1460px){.xju9v9y.xju9v9y{color:blue}}",
+    priority: 3200
+});
+export const styles = {
+    a: {
+        k1xSpc: "x1s85apg x1gcnmh1",
+        $$css: true
+    },
+    b: {
+        kMwMTN: "x1e2nbdu xju9v9y",
+        $$css: true
+    }
+};
