+++ crates/stylex-transform/tests/__swc_snapshots__/tests/transform_stylex_create_test/media_query_canonicalization.rs/media_query_order_opt_out_keeps_queries_verbatim.js
@@ -1,0 +1,29 @@
+import _inject from "@stylexjs/stylex/lib/stylex-inject";
+var _inject2 = _inject;
+import * as stylex from '@stylexjs/stylex';
+_inject2({
+    ltr: ".x1s85apg{display:none}",
+    priority: 3000
+});
+_inject2({
+    ltr: "@media (max-height:120px) and (min-width: 720px){.x4ob7n2.x4ob7n2{display:block}}",
+    priority: 3200
+});
+_inject2({
+    ltr: ".x1e2nbdu{color:red}",
+    priority: 3000
+});
+_inject2({
+    ltr: "@media (width >= 1460px){.xy2bn39.xy2bn39{color:blue}}",
+    priority: 3200
+});
+export const styles = {
+    a: {
+        k1xSpc: "x1s85apg x4ob7n2",
+        $$css: true
+    },
+    b: {
+        kMwMTN: "x1e2nbdu xy2bn39",
+        $$css: true
+    }
+};
