+++ crates/stylex-transform/tests/transform_stylex_create_test/media_query_canonicalization.rs
@@ -21,7 +21,7 @@
 use crate::utils::prelude::*;
 
 /// The reproduction from issue #1250, verbatim.
-const ISSUE_1250_INPUT: &str = r#"
+const INPUT_CODE: &str = r#"
     import * as stylex from '@stylexjs/stylex';
     export const styles = stylex.create({
       a: {
@@ -39,78 +39,13 @@ const ISSUE_1250_INPUT: &str = r#"
     });
   "#;
 
-stylex_test_transform!(
-  authored_media_queries_are_canonicalized,
-  ISSUE_1250_INPUT,
-  r#"
-    import _inject from "@stylexjs/stylex/lib/stylex-inject";
-    var _inject2 = _inject;
-    import * as stylex from '@stylexjs/stylex';
-    _inject2({
-      ltr: ".x1s85apg{display:none}",
-      priority: 3000
-    });
-    _inject2({
-      ltr: "@media (min-width: 720px) and (max-height: 120px){.x1gcnmh1.x1gcnmh1{display:block}}",
-      priority: 3200
-    });
-    _inject2({
-      ltr: ".x1e2nbdu{color:red}",
-      priority: 3000
-    });
-    _inject2({
-      ltr: "@media (min-width: 1460px){.xju9v9y.xju9v9y{color:blue}}",
-      priority: 3200
-    });
-    export const styles = {
-      a: {
-        k1xSpc: "x1s85apg x1gcnmh1",
-        $$css: true
-      },
-      b: {
-        kMwMTN: "x1e2nbdu xju9v9y",
-        $$css: true
-      }
-    };
-  "#
-);
+stylex_test!(authored_media_queries_are_canonicalized, INPUT_CODE);
 
-stylex_test_transform!(
+stylex_test!(
   media_query_order_opt_out_keeps_queries_verbatim,
   |tr| build_test_transform(tr.comments.clone(), |b| {
     b.with_enable_media_query_order(false)
       .with_runtime_injection()
   }),
-  ISSUE_1250_INPUT,
-  r#"
-    import _inject from "@stylexjs/stylex/lib/stylex-inject";
-    var _inject2 = _inject;
-    import * as stylex from '@stylexjs/stylex';
-    _inject2({
-      ltr: ".x1s85apg{display:none}",
-      priority: 3000
-    });
-    _inject2({
-      ltr: "@media (max-height:120px) and (min-width: 720px){.x4ob7n2.x4ob7n2{display:block}}",
-      priority: 3200
-    });
-    _inject2({
-      ltr: ".x1e2nbdu{color:red}",
-      priority: 3000
-    });
-    _inject2({
-      ltr: "@media (width >= 1460px){.xy2bn39.xy2bn39{color:blue}}",
-      priority: 3200
-    });
-    export const styles = {
-      a: {
-        k1xSpc: "x1s85apg x4ob7n2",
-        $$css: true
-      },
-      b: {
-        kMwMTN: "x1e2nbdu xy2bn39",
-        $$css: true
-      }
-    };
-  "#
+  INPUT_CODE
 );
